backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aerich>=0.9.2",
    "bcrypt>=5.0.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.0",
    "gitpython>=3.1.46",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.6",
    "openai>=2.15.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "tortoise-orm[asyncpg]>=0.25.3",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]

[tool.aerich]
tortoise_orm = "core.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
